+++ Vanilla/Vanilla.Character.Database/character_spell_cooldown.cs
@@ -17,6 +17,6 @@ namespace Vanilla.Character.Database
         public long guid { get; set; }
         public long spell { get; set; }
         public long item { get; set; }
-        public long time { get; set; }
+        public decimal time { get; set; }
     }
 }
