+++ Vanilla/Vanilla.World.Database/creature_onkill_reputation.cs
@@ -17,11 +17,11 @@ namespace Vanilla.World.Database
         public int creature_id { get; set; }
         public short RewOnKillRepFaction1 { get; set; }
         public short RewOnKillRepFaction2 { get; set; }
-        public short MaxStanding1 { get; set; }
-        public short IsTeamAward1 { get; set; }
+        public sbyte MaxStanding1 { get; set; }
+        public sbyte IsTeamAward1 { get; set; }
         public int RewOnKillRepValue1 { get; set; }
-        public short MaxStanding2 { get; set; }
-        public short IsTeamAward2 { get; set; }
+        public sbyte MaxStanding2 { get; set; }
+        public sbyte IsTeamAward2 { get; set; }
         public int RewOnKillRepValue2 { get; set; }
         public byte TeamDependent { get; set; }
     }
